Split run/show.go into show.go, watch.go, and github.go
